Extract downsampling methods
